#1047 P1: relocate session.rs inline tests to session_tests.rs
session.rs went from 3,108 raw LOC (1,352 prod / 1,756 tests) to 1,355
by moving the entire `#[cfg(test)] mod tests { ... }` block into a
sibling `session_tests.rs` file, loaded via
`#[cfg(test)] #[path = "session_tests.rs"] mod tests;` appended to
session.rs.

Pure file-organization split — no test bodies modified, no production
code touched. The test module's `super::*` continues to resolve into
session.rs because the module path `crate::session::tests` is
unchanged; `#[path]` only redirects the file location.

LOC distribution after:
- session.rs:        1,355 (was 3,108) — under threshold
- session_tests.rs:  1,757 (new — same content, dedented one level)

cargo test --release: 865 passed, 0 failed.

Closes the LOC-driven portion of #1047. Further structural splits
(key/table/wheel/entry submodules) are deferred — session.rs is now
under the modularity threshold and the structural split is a
code-organization win rather than a discipline mandate.
